+++ js/ghu-install.js
@@ -12,7 +12,7 @@ jQuery( document ).ready(
 	function ($) {
 		// Hide non-default (Bitbucket & GitLab) settings on page load.
 		$.each(
-			['bitbucket', 'gitlab', 'gitea'], function () {
+			['bitbucket', 'gitlab', 'gitea', 'zipfile'], function () {
 				$( 'input.'.concat( this, '_setting' ) ).parents( 'tr' ).hide();
 			}
 		);
@@ -22,7 +22,7 @@ jQuery( document ).ready(
 			'change', function () {
 
 				// create difference array.
-				var hideMe = $( ['github', 'bitbucket', 'gitlab', 'gitea'] ).not( [this.value] ).get();
+				var hideMe = $( ['github', 'bitbucket', 'gitlab', 'gitea', 'zipfile'] ).not( [this.value] ).get();
 
 				/*
 				* Show/hide all settings that have the selected api's class.